requirements.txt
pydantic==2.7.4
httpx==0.27.0
playwright==1.48.0
rich==13.9.2
ujson==5.10.0
python-json-logger==2.0.7
pyyaml==6.0.2
